+++ content/umetaflow_results_old.py
@@ -1,6 +1,6 @@
 import streamlit as st
 from src.common.common import page_setup
-from src.UmetaFlowTOPPWorkflowNewResults import Workflow
+from src.UmetaFlowTOPPWorkflowOld import Workflow
 
 # The rest of the page can, but does not have to be changed
 params = page_setup()
